refactor: deliver sprint1 wave3 service extraction (#538)
