+++ openspec/_ops/task_runs/ISSUE-538.md
@@ -1,0 +1,132 @@
+# ISSUE-538
+
+- Issue: #538
+- Issue URL: https://github.com/Leeky1017/CreoNow/issues/538
+- Branch: task/538-s1-wave3-governed-delivery
+- PR: 待回填（由 PR 创建步骤更新为真实链接）
+- Scope:
+  - `s1-doc-service-extract`
+  - `s1-ai-service-extract`
+  - `s1-kg-service-extract`
+  - `openspec/changes/EXECUTION_ORDER.md`
+  - `openspec/_ops/task_runs/ISSUE-538.md`
+- Out of Scope:
+  - 非 Wave 3 的需求扩展
+  - 与服务提取无关的 UI/交互改动
+
+## Plan
+
+- [x] 创建 OPEN issue 与隔离 worktree
+- [x] 建立 Rulebook 任务并 validate
+- [x] 主会话并行派发 3 路子代理实现 Wave 3 changes
+- [x] 主会话审计子代理结果并修正偏差
+- [x] 完成 Wave 3 代码实现与回归验证
+- [ ] preflight / PR / auto-merge / main 同步 / worktree 清理
+
+## Runs
+
+### 2026-02-14 18:17-18:18 任务准入与环境隔离
+
+- Command:
+  - `git fetch origin main && git merge --ff-only origin/main`
+  - `gh issue create --title "Deliver Sprint1 Wave3 changes with governed subagent execution" --body-file /tmp/issue_wave3_delivery.md`
+  - `scripts/agent_worktree_setup.sh 538 s1-wave3-governed-delivery`
+  - `pnpm install --frozen-lockfile`
+  - `rulebook task create issue-538-s1-wave3-governed-delivery`
+  - `rulebook task validate issue-538-s1-wave3-governed-delivery`
+- Exit code: `0`
+- Key output:
+  - Issue 创建成功：`#538`
+  - Worktree：`.worktrees/issue-538-s1-wave3-governed-delivery`
+  - Branch：`task/538-s1-wave3-governed-delivery`
+  - 依赖安装成功，Rulebook task 校验通过
+
+### 2026-02-14 18:20-18:31 子代理并行实现（Wave 3）
+
+- Sub-agent sessions:
+  - `019c5baa-7bd5-7813-83c7-aa9660d38ad9` → `s1-doc-service-extract`
+  - `019c5baa-7bec-7c41-a27b-ec82219b27f4` → `s1-ai-service-extract`
+  - `019c5baa-7c0f-7443-9123-69fb51220843` → `s1-kg-service-extract`
+- Exit code: `0`（3 路会话均返回）
+- Key output:
+  - `s1-doc-service-extract`：新增 `documentCoreService/types/documentCrud/version/branch`，`documentService.ts` 收敛为 facade，补齐 `S1-DSE-S1/S2/S3` 测试
+  - `s1-ai-service-extract`：提取 `runtimeConfig/errorMapper/providerResolver`，`aiService.ts` 改为委托，补齐 `AI-S1-ASE-S1/S2/S3` 测试
+  - `s1-kg-service-extract`：新增 `kgCoreService/types/kgQuery/kgWrite`，`kgService.ts` 改为 facade，补齐 `KG-S1-KSE-S1/S2/S3` 测试
+
+### 2026-02-14 18:32-18:38 主会话审计与复测（映射 + 回归）
+
+- Command:
+  - 文档服务链路：
+    - `pnpm exec tsx apps/desktop/main/src/services/documents/__tests__/document-service-extract.structure.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/documents/__tests__/document-service-no-duplicate-implementation.test.ts`
+    - `pnpm exec tsx apps/desktop/tests/integration/document-management/document-service-facade-contract.test.ts`
+    - `pnpm exec tsx apps/desktop/tests/unit/documentService.lifecycle.test.ts`
+    - `pnpm exec tsx apps/desktop/tests/unit/document-ipc-contract.test.ts`
+    - `pnpm exec tsx apps/desktop/tests/unit/version-hardening-boundary.ipc.test.ts`
+  - AI 服务链路：
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/ai-runtime-and-error-extract.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/providerResolver-state-isolation.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/ai-public-contract-regression.test.ts`
+    - `pnpm exec tsx apps/desktop/tests/unit/ai-upstream-error-mapping.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/provider-failover-half-open.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/aiService-provider-unavailable.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/llm-proxy-retry-rate-limit.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/ai/__tests__/aiService-runtime-multiturn.test.ts`
+  - KG/Context 链路：
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/kg-service-split-boundary.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/kg-service-facade-delegation.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/kg-service-exports-visibility.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/kgService.contextLevel.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/kgService.aliases.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/kg/__tests__/entityMatcher.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/context/__tests__/rulesFetcher.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/context/__tests__/retrievedFetcher.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/context/__tests__/formatEntity.import-boundary.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/context/__tests__/layerAssemblyService.contract-regression.test.ts`
+    - `pnpm exec tsx apps/desktop/main/src/services/context/__tests__/layerAssemblyService.dependency-graph.test.ts`
+- Exit code: `0`
+- Key output:
+  - Wave 3 场景映射测试与受影响回归全部通过
+  - `documentService.lifecycle`、`version-hardening-boundary`、`ai-upstream-error-mapping`、`kg/context` 关键链路保持绿灯
+
+### 2026-02-14 18:39-18:41 主会话纠偏与工程门禁
+
+- Command:
+  - `LC_ALL=C sed -i '1s/^\xEF\xBB\xBF//' apps/desktop/main/src/services/ai/aiService.ts`
+  - 去重 `documentCoreService.ts` 类型来源，改为引用 `types.ts`（主会话审计修复）
+  - `pnpm typecheck && pnpm lint && pnpm cross-module:check && pnpm test:unit`
+- Exit code: `0`
+- Key output:
+  - 清除 `aiService.ts` BOM 非 ASCII 噪声
+  - 修复 `documentCoreService.ts` 类型去重引入的 TS6196 问题
+  - 工程门禁全绿（`lint` 仅历史 warning，无 error）
+
+### 2026-02-14 18:41 治理收口（change 归档 + 执行顺序同步）
+
+- Command:
+  - `sed -i 's/- \[ \]/- [x]/g' openspec/changes/{s1-doc-service-extract,s1-ai-service-extract,s1-kg-service-extract}/tasks.md`
+  - `mv openspec/changes/{s1-doc-service-extract,s1-ai-service-extract,s1-kg-service-extract} openspec/changes/archive/`
+  - 更新 `openspec/changes/EXECUTION_ORDER.md`（活跃 change=0）
+- Exit code: `0`
+- Key output:
+  - Wave 3 三个 change 已归档
+  - `EXECUTION_ORDER.md` 已同步为“无活跃 change”状态
+
+## Dependency Sync Check
+
+- Inputs:
+  - `docs/plans/unified-roadmap.md`（Wave 3 定义与执行建议）
+  - `openspec/specs/document-management/spec.md`
+  - `openspec/specs/ai-service/spec.md`
+  - `openspec/specs/knowledge-graph/spec.md`
+  - `openspec/changes/archive/{s1-doc-service-extract,s1-ai-service-extract,s1-kg-service-extract}/proposal.md`
+- Result:
+  - `s1-doc-service-extract`: `NO_DRIFT`
+  - `s1-ai-service-extract`: `NO_DRIFT`
+  - `s1-kg-service-extract`: `NO_DRIFT`
+- Reason:
+  - 三条 change 均限定在“服务提取 + 门面契约保持 + 旧实现去重”边界内，未扩展新业务能力或修改外部 IPC 契约。
+
+## Main Session Audit
+
+- 待补充（PR 前补齐签字字段，最终提交仅修改本 RUN_LOG）
