+++ openspec/changes/archive/s1-ai-service-extract/specs/ai-service-delta.md
