+++ openspec/changes/archive/s1-doc-service-extract/proposal.md
