+++ openspec/changes/archive/s1-doc-service-extract/specs/document-management-delta.md
