+++ openspec/changes/archive/s1-kg-service-extract/proposal.md
