+++ openspec/changes/archive/s1-kg-service-extract/specs/knowledge-graph-delta.md
