+++ rulebook/tasks/issue-538-s1-wave3-governed-delivery/.metadata.json
@@ -1,0 +1,5 @@
+{
+  "status": "pending",
+  "createdAt": "2026-02-14T10:18:15.553Z",
+  "updatedAt": "2026-02-14T10:18:15.553Z"
+}
