+++ rulebook/tasks/issue-538-s1-wave3-governed-delivery/specs/governance/spec.md
@@ -1,0 +1,15 @@
+# Governance Task Spec — issue-538-s1-wave3-governed-delivery
+
+## Scope
+
+Deliver Sprint 1 Wave 3 changes under OpenSpec + Rulebook + GitHub governance:
+
+- `s1-doc-service-extract`
+- `s1-ai-service-extract`
+- `s1-kg-service-extract`
+
+## Acceptance
+
+- All three Wave 3 changes are implemented and verified against their delta specs/tasks.
+- Required checks pass with auto-merge enabled.
+- Delivery closes on control-plane `main` with RUN_LOG and main-session audit evidence.
